handling customer DELETE not found exception

+++ src/main/java/com/dwanford/spring5mvcrest/services/CustomerServiceImpl.java
@@ -67,6 +67,8 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void deleteCustomerById(Long id) {
-        customerRepository.deleteById(id);
+        if(customerRepository.findById(id).isPresent())
+            customerRepository.deleteById(id);
+        else throw new ResourceNotFoundException();
     }
 }
